--- src/blocks/settings/domain/index.ts
@@ -1,2 +1,0 @@
-// Settings domain exports
-export * from './dynamicFormAccess';
