+++ src/blocks/settings/index.ts
@@ -1,5 +1,5 @@
-export { default as VisibilityForm } from './VisibilityForm';
 export { default as StyleForm } from './StyleForm';
-export { default as DynamicPropsForm } from './DynamicPropsForm';
+export { default as BlockPropertiesForm } from './BlockPropertiesForm';
+export { default as VisibilityForm } from './VisibilityForm';
 export type { BlockFormProps } from '@blocks/shared/FormTypes';
 export type { Field } from '@blocks/shared/Field';
