--- src/renderer/domain/index.ts
@@ -1,2 +1,0 @@
-// Renderer domain exports
-export * from './visibilityRules';
